+++ common/src/main/java/mod/adrenix/nostalgic/common/config/DefaultConfig.java
@@ -85,6 +85,7 @@ public abstract class DefaultConfig
         public static final boolean REMOVE_TITLE_MOD_LOADER_TEXT = true;
         public static final boolean REMOVE_TITLE_ACCESSIBILITY = false;
         public static final boolean REMOVE_TITLE_LANGUAGE = false;
+        public static final boolean REMOVE_TITLE_REALMS = false;
         public static final boolean OVERRIDE_TITLE_SCREEN = true;
         public static final boolean OLD_TITLE_BACKGROUND = true;
         public static final boolean OLD_ALPHA_LOGO = true;